Correction de install_hotspot pour tenir compte de la diversité des wifi devices

+++ ConfigTinker/install_hotspot.sh
@@ -10,25 +10,33 @@ moi=$(whoami)
 ######
 # Get mac_address
 ######
-mac_address=$(sudo cat /sys/class/net/wlan0/address)
+# get wifi device
+device=$(basename $(find /sys/class/net -name wl*))
+if test ! -z $device
+then
+	mac_address=$(sudo cat /sys/class/net/$device/address)
 ######
 # Créer l'uuid
 ######
-hsuuid=$(uuid)
+	hsuuid=$(uuid)
 ######
 # Création du hotspot
 ######
-hotspot_name="$moi""box"
-fic1=$(tempfile)
-cat nafabox.template | sed -e "s/MOI/${moi}/g" > $fic1
-fic2=$(tempfile)
-cat $fic1 | sed -e "s/UUID/${hsuuid}/g" > $fic2
-fic3=$(tempfile)
-cat $fic2 | sed -e "s/MAC_ADDRESS/${mac_address}/g" > $fic3
-sudo cp $fic3 /etc/NetworkManager/system-connections/$hotspot_name
-sudo systemctl stop NetworkManager
-sleep 5
-sudo systemctl start NetworkManager
-
+	hotspot_name="$moi""box"
+	fic0=$(tempfile)
+	cat nafabox.template | sed -e "s/wlan0/${device}/g" > $fic0
+	fic1=$(tempfile)
+	cat $fic0 | sed -e "s/MOI/${moi}/g" > $fic1
+	fic2=$(tempfile)
+	cat $fic1 | sed -e "s/UUID/${hsuuid}/g" > $fic2
+	fic3=$(tempfile)
+	cat $fic2 | sed -e "s/MAC_ADDRESS/${mac_address}/g" > $fic3
+	sudo cp $fic3 /etc/NetworkManager/system-connections/$hotspot_name
+	sudo systemctl stop NetworkManager
+	sleep 5
+	sudo systemctl start NetworkManager
+else
+	echo "No wifi decice found"
+fi
